지은/0805/왜 안될까..댓글

+++ 404_mid_project/src/main/java/model/post/PostDto.java
@@ -2,7 +2,7 @@ package model.post;
 
 public class PostDto {
 	private int postNum;
-	private int postWriterNum;
+	private Long postWriterNum;
 	private String postTitle;
 	private String postContent;
 	private int postType;
@@ -18,10 +18,10 @@ public class PostDto {
 	public void setPostNum(int postNum) {
 		this.postNum = postNum;
 	}
-	public int getPostWriterNum() {
+	public Long getPostWriterNum() {
 		return postWriterNum;
 	}
-	public void setPostWriterNum(int postWriterNum) {
+	public void setPostWriterNum(Long postWriterNum) {
 		this.postWriterNum = postWriterNum;
 	}
 	public String getPostTitle() {
